requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
django-imagekit
djangorestframework-simplejwt